Updated a slightly better background image
Created private class RoadDash inside SimulationArea to instantiate road dashes.

Adding a tick in SimulationWorker to time events/animations

Implemented a simple seconds counter that can be used for timing and displaying the time elapsed in the simulation so far.

Added methods to initialize a number of dashes on the road at the start of the simulation.

Added methods to animate dashes across the screen.

Next steps: Remove dashes that have made across the screen, instantiate new dashes at the right side of screen at an interval dependent on the speed of the car.