Added methods to instantiate new dashes and translate them across the road.
Removed initial dashes and replaced them with a starting line.

Simulation area takes the actual speed setting from dashboard and adjusts the speed of animation appropriately.

Added two black side panels to hide the dashes that are finished but not yet removed as well as the dashes that are being instantiated on the other side.

Next steps:

Get a car and edit rpm of wheels to simulate speed.

Add acceleration values to create an interpolation on the animation.

Constraints: Right now the window is resizable which resizes the background image. It would make things much easier to lock the size of the window.