Fixed bug that made the car appear to move backwards when the brake is engaged.
The dashes were moving about 60x faster than they were supposed to, which gave the illusion of proper behavior because of the dashes blurring together. Then when the car would decelerate, the actual shape/speed of the dashes is perceived and gives the illusion of the car moving backwards rapidly.

Once I fixed the speed to calculate meters/ 16 ms, it worked.